feat(GenericContainer): add PullPolicySetting trait for managing image pull policies
