+++ tests/Unit/Containers/GenericContainer/PullPolicySettingTest.php
@@ -1,0 +1,43 @@
+<?php
+
+/** @noinspection PhpUnhandledExceptionInspection */
+
+namespace Tests\Unit\Containers\GenericContainer;
+
+use PHPUnit\Framework\TestCase;
+use Testcontainers\Containers\GenericContainer\GenericContainer;
+use Testcontainers\Containers\GenericContainer\PullPolicySetting;
+use Testcontainers\Containers\ImagePullPolicy;
+
+class PullPolicySettingTest extends TestCase
+{
+    public function testHasPullPolicySettingTrait()
+    {
+        $uses = class_uses(GenericContainer::class);
+
+        $this->assertContains(PullPolicySetting::class, $uses);
+    }
+
+    public function testStaticPullPolicy()
+    {
+        $container = new PullPolicySettingWithStaticPullPolicyContainer('alpine:latest');
+        $instance = $container->start();
+
+        $this->assertSame(ImagePullPolicy::$ALWAYS, $instance->getImagePullPolicy()->toString());
+    }
+
+    public function testStartWithImagePullPolicy()
+    {
+        $container = (new GenericContainer('alpine:latest'))
+            ->withImagePullPolicy(ImagePullPolicy::MISSING());
+        /** @noinspection PhpUnhandledExceptionInspection */
+        $instance = $container->start();
+
+        $this->assertSame(ImagePullPolicy::$MISSING, $instance->getImagePullPolicy()->toString());
+    }
+}
+
+class PullPolicySettingWithStaticPullPolicyContainer extends GenericContainer
+{
+    protected static $PULL_POLICY = 'always';
+}
